refactor(mcp-utils): Extract ToolFilter into mcp-utils
